+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # CHANGELOG
 
+### 3.0.0
+
+* All assets are now requested cache-bust query parameter (`__uncache=${ Date.now() }`)
+* Matching assets in cache now uses `ignoreSearch` search option of `CacheQueryOptions`, with [sw-cache-options](https://github.com/NekR/sw-cache-options) polyfill for Chrome.
+* Rename `scope` option to `publicPath` (`scope` is deprecated now and will produce warnings upon use)
+* Make `publicPath` `''` (empty string) by default
+* Make `relativePaths` `true` by default
+* Cache sections `'additional'` and `'optional'` are now allowed only when `updateStrategy`option is set to `'changed'`
+* `updateStrategy: `changed`` now uses `version` option instead of compilation hash.
+* Version now is not set by default and returns compilation hash for `updateStrategy: 'changed`` and `version` for `updateStrategy: 'all'` (when not set)
+* Added `alwaysRevalidate` option
+* Improve `updateStrategy` option implementation. `changed` mode might be still unstable--**do not use it yet**
+* `install()` method signature now is `install(options)` (callbacks are removed)
+* Runtime events are not implemented for ServiceWorker: `onUpdating`, `onUpdateReady`, `onUpdated`, `onError`.  
+  Example: `runtime.install({ onError: () => ... })`
+* Added `applyUpdate()` method of runtime
+* Absolute URLs can now be specified in `caches` as any other assets (they are required to be marked as `externals`)
+
 ### 2.1.0
 
 * Allow pattern matching in `externals`